comparisonrecipeselect: fix logic for setting modal title
If you clicked on "Select" then cancelled the modal that appears, then
clicked on "No match" then the title of the modal in the second instance
retains the recipe name from the first time which is wrong. Rearrange
the logic so that this cannot happen (and make it tidier at the same
time).

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>