update: fix commands.getstatusoutput errors when parsing python2 branch
If you update a python2-based branch then the python2-compatible version
of bitbake will be checked out, but we are calling into bitbake's
bb.utils directly here from python 3 and thus you get an error about
commands.getstatusoutput being missing (since that is not available in
python 3 and the old version of bitbake refers to it). To fix this,
check out origin/master in the bitbake repo right before we call the
code in question.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>