Add some text to the password reset email
What we had before was a little bit terse, so add some reasonable text.

Also mention in the confirmation page that sending an email is
predicated on there actually being an account matching the specified
email address (and we deliberately don't specify whether there is or
not, in order to prevent user enumeration).

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>