Fix Submit/Edit layer form field rendering issue
Some controls on the Submit / Edit Layer page were getting the wrong
width and any "feedback" elements were being pulled to the right because
of the size of their parent divs. Instead of pulling in the size of the
form control, act on the div to fix the problem. Additionally the
help text should now be allowed to expand to the full width of the
parent div since that is now sensible.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>