dockersetup: obtain and use the real project name instead of hardcoding
We need the real project name here otherwise we are mounting something
nonexistent instead of the real volume we want to populate. This fixes
a bug where the static assets were not available (thus the web interface
looks awful) if the directory you installed the application from was not
called "layerindex-web".

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>