RRS: skip problematic OE-Core commits (when a dependency)
There are a range of commits in OE-Core which cause parsing problems;
map them to the one that fixes it in order to avoid the problem. (This
will only be done if we're dealing with OE-Core as a dependency, rather
than the actual layer we're parsing).

(The second set are some commits during the python 3 conversion time.)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>