RRS: ensure default URLs for release/milestone are the latest
In the "Maintenance Plan" drop-down the maintenance plans point to the
"default" release and milestone, but it was picking the most recently
added record in the database rather than the latest one by date. Use an
order_by() to ensure we get the most recent release/milestone by date
rather than just the most recently added in case they have been added
out-of-order.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>