RRS: Handle two versions added on same day then later one deleted
We have at least one instance where two versions of a recipe were added
at the same time and then later one was deleted - sed. We didn't detect
more than one recipe being added and thus the delete was seen as
removing the recipe entirely, causing the recipe to vanish. Fix the
filter so that we see the other addition and adjust the debug printing
so that we can see what type of deletions are occurring.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>