rrs_upstream_history: fix get_upstream_info
'result' is a list of RecipeUpstream objects... drop legacy 'recipe, ru'
syntax. In the current code base, result.append((recipe, ru)) was
appending a tuple of the last value of 'recipe' (in 'for recipe_qry:' under
'for maintplan in maintplans:') and the RecipeUpstream object.

Fixes:
AttributeError: 'tuple' object has no attribute 'history'

Signed-off-by: Tim Orling <tim.orling@konsulko.com>