Configure ruff code formater to preserve quote style
Current code uses a mix of single and double quotes, and that causes a
lot of code changes every time a file is formatted with ruff. This ruff
configuration gets us all the benfits of automated code formating
withough the downside of having to unify the whole code base in terms of
quote styling.

Signed-off-by: Piotr Buliński <piotr@qbee.io>
Signed-off-by: Tim Orling <tim.orling@konsulko.com>